# RFE: encoder: add stride & various raw input formats support