--- a/tests/test_misc.rs
+++ b/tests/test_misc.rs
@@ -4,7 +4,7 @@ use qoi::{
 };
 
 #[test]
-fn test_new_encoder() {
+fn test_new_decoder() {
     // this used to fail due to `Bytes` not being `pub`
     let arr = [0u8];
     let _ = qoi::Decoder::new(&arr[..]);