@elmarco 2025-02-13
See also: https://github.com/Devolutions/IronRDP/pull/670