@elmarco 2025-03-18
@aldanor hi, wdyt?