@elmarco 2025-05-27
> > You need alpha channel input for alpha channel output.
> 
> Adding an alpha channel with default values of 255 would be imho fine and not unexpected.

Well, this use case goes beyond the typical simple encoding imo.