@Joshix-1 2025-05-28
Why is `RawChannels::bytes_per_pixel` not public? There should either be another Encoder constructor without a stride argument or `bytes_per_pixel` should be public. Otherwise the `Encoder::new_raw` API is imho too clunky to use if you have no stride.