@aldanor 2025-07-28
>  I am trying to understand where it comes from and how to fix it

It might come from a double loop with now-unknown number of iterations in each preventing some common loop optimizations? I guess you can godbolt something equivalent.

Yea, -33% is pretty bad for sure... -3% would still be not nice but within noise bounds.