@elmarco 2025-07-28
> Yea, -33% is pretty bad for sure... -3% would still be not nice but within noise bounds.

it turns out it's the assert_eq!(), this adds a bunch of panic handling code and prevent some optimizations. I switched it to debug_assert!(). Now, no performance regression is observable (with perf stat).