@Joshix-1 2025-07-28
I ran `cargo run --release ../assets/` in the bench folder