@Joshix-1 2025-07-28
[ce61683e8682edb68f8f040f4cbbce1eae143fb7 ](https://github.com/Joshix-1/qoi-rust/commit/ce61683e8682edb68f8f040f4cbbce1eae143fb7) has a similar effect. Seems to be some weird case where the compiler doesn't optimize