@Joshix-1 2025-07-28
Created a [flamegraph](https://github.com/flamegraph-rs/flamegraph) and saw Try take a big chunk.
<img width="1920" height="675" alt="grafik" src="https://github.com/user-attachments/assets/b6dd8e31-08e9-4a87-8e3e-cd9129bed940" />