@elmarco 2025-07-29
I tried to make a subset test case to report to the compiler without success atm.

@Joshix-1 what to do next?:
- use your workaround
- add the feature flag I proposed
- look for other solution (is stream writer really useful?)
- wait for a compiler fix

thanks