@Joshix-1 2025-07-29
Could not improve performance of master. I would suggest using https://github.com/aldanor/qoi-rust/commit/50293f36d300511bbd5b16a6a1bb646ee63b98b8

With 
```toml
[profile.release]
opt-level = 3
debug = true
```
in `bench/Cargo.toml` I get

```
$ cargo run --release ../assets/
Overall results: (7 images, 7.78 MB raw, 2.48 MP):
--------------------------------------
                     qoi.h    qoi-rust
--------------------------------------
decode   Mp/s        274.2       383.2
         MB/s        859.3      1200.9
--------------------------------------
encode   Mp/s        220.7       260.4
         MB/s        691.7       816.0
--------------------------------------
```

without the `debug = true` I get

```
Overall results: (7 images, 7.78 MB raw, 2.48 MP):
--------------------------------------
                     qoi.h    qoi-rust
--------------------------------------
decode   Mp/s        280.9       370.7
         MB/s        880.2      1161.4
--------------------------------------
encode   Mp/s        226.5       252.6
         MB/s        709.8       791.6
--------------------------------------
```

on [master](https://github.com/aldanor/qoi-rust/commit/81c14c4b637c326a801e40701460b654d46e8fbe) I get:

```
Overall results: (7 images, 7.78 MB raw, 2.48 MP):
--------------------------------------
                     qoi.h    qoi-rust
--------------------------------------
decode   Mp/s        284.8       368.8
         MB/s        892.4      1155.7
--------------------------------------
encode   Mp/s        220.2       257.1
         MB/s        690.0       805.7
--------------------------------------
```

But I don't really know anymore, it's all really weird and unexpected