@elmarco 2025-07-29
indeed, I am ok with https://github.com/aldanor/qoi-rust/commit/50293f36d300511bbd5b16a6a1bb646ee63b98b8, we can later revert it.