@Joshix-1 2025-07-29
Why revert it later? The change makes sense. Infallible methods shouldn't return results that could contain errors